Relocate appLayout, header and footer files
